video-translator/scripts/requirements.txt
openai-whisper
edge-tts
ffmpeg-python
tqdm